+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 - remove reference to vendored library ``django.utils.six`` in favor of ``six`` (gh-526)
 - Fixed hardcoded history manager (gh-540)
 
+Unreleased
+----------
+- Added the possibility to create a relation to the original model
+
 2.7.0 (2019-01-16)
 ------------------
 - Add support for ``using`` chained manager method and save/delete keyword argument (gh-507)